+++ lib/src/emitter.dart
@@ -64,6 +64,15 @@ class DartEmitter extends Object
     return new DartEmitter(new Allocator.simplePrefixing());
   }
 
+  /// Whether the provided [method] is considered a lambda method.
+  static bool _isLambdaMethod(Method method) =>
+      method.lambda ?? method.body is ToCodeExpression;
+
+  /// Whether the provided [constructor] is considered a lambda method.
+  static bool _isLambdaConstructor(Constructor constructor) =>
+      constructor.lambda ??
+      constructor.factory && constructor.body is ToCodeExpression;
+
   @override
   visitAnnotation(Expression spec, [StringSink output]) {
     (output ??= new StringBuffer()).write('@');
@@ -112,7 +121,7 @@ class DartEmitter extends Object
     });
     spec.methods.forEach((m) {
       visitMethod(m, output);
-      if (m.lambda) {
+      if (_isLambdaMethod(m)) {
         output.write(';');
       }
       output.writeln();
@@ -189,7 +198,7 @@ class DartEmitter extends Object
       visitType(spec.redirect.type, output);
       output.write(';');
     } else if (spec.body != null) {
-      if (spec.lambda) {
+      if (_isLambdaConstructor(spec)) {
         output.write(' => ');
         spec.body.accept(this, output);
         output.write(';');
@@ -279,7 +288,7 @@ class DartEmitter extends Object
     final body = new StringBuffer();
     for (final spec in spec.body) {
       spec.accept(this, body);
-      if (spec is Method && spec.lambda) {
+      if (spec is Method && _isLambdaMethod(spec)) {
         body.write(';');
       }
     }
@@ -408,13 +417,13 @@ class DartEmitter extends Object
             break;
         }
       }
-      if (spec.lambda) {
+      if (_isLambdaMethod(spec)) {
         output.write(' => ');
       } else {
         output.write(' { ');
       }
       spec.body.accept(this, output);
-      if (!spec.lambda) {
+      if (!_isLambdaMethod(spec)) {
         output.write(' } ');
       }
     } else {